pyproject.toml
[project]
name = "securecode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "minio>=7.2.15",
    "python-dotenv>=1.1.1",
]
